+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Change Log
 
-### Unreleased
+### [v6.190](https://github.com/replicant4j/replicant/tree/v6.190) (2026-05-19) · [Full Changelog](https://github.com/replicant4j/replicant/compare/v6.189...v6.190)
+
+Changes in this release:
 
 * Change some `java.util.List` to `java.util.Collection` in `AbstractSessionContextImpl`
 * Change `ReplicantSession.isSubscriptionEntryPresent()` to be public